change elements, and style, add grid
